Journey app
This my Journey app created on this March. And this  application is for
trekking. User can know which place is awesome to trek and travel. And
also this is also a tutorial to learn the code which use table view
controller, collection view, core data, how to add icon and plot route
with map kit.